requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
django-cors-headers==4.5.0
django-filter==24.3
djangorestframework==3.15.2
drf-yasg==1.21.8
exceptiongroup==1.2.2
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tomli==2.0.2
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
